refactor: extract GeminiOrchestrator to modular structure
- Move GeminiOrchestrator class from main.py to src/orchestrators/gemini_orchestrator.py
- Reduce main.py from 1232 to 808 lines for better maintainability
- Create proper module structure with __init__.py for orchestrators
- Import GeminiOrchestrator from new location in main.py
- All API endpoints tested and confirmed working after refactoring

This improves code organization and makes the codebase more maintainable.
